requirements.txt
babel
python-dateutil
flask-moment
flask-wtf
flask_sqlalchemy
psycopg2-binary
Flask-Migrate